Refactor according to review
