+++ offer-manager-app/src/main/groovy/life/qbic/portal/offermanager/DependencyManager.groovy
@@ -18,28 +18,28 @@ import life.qbic.portal.offermanager.dataresources.products.ProductsDbConnector
 import life.qbic.portal.offermanager.dataresources.products.ProductsResourcesService
 import life.qbic.portal.offermanager.dataresources.offers.OfferUpdateService
 import life.qbic.portal.offermanager.dataresources.database.OverviewService
-import life.qbic.portal.offermanager.components.createaffiliation.CreateAffiliationController
-import life.qbic.portal.offermanager.components.createperson.CreatePersonController
-import life.qbic.portal.offermanager.components.createoffer.CreateOfferController
+import life.qbic.portal.offermanager.components.affiliation.create.CreateAffiliationController
+import life.qbic.portal.offermanager.components.person.create.CreatePersonController
+import life.qbic.portal.offermanager.components.offer.create.CreateOfferController
 import life.qbic.portal.offermanager.web.controllers.ListAffiliationsController
 import life.qbic.portal.offermanager.web.controllers.SearchCustomerController
-import life.qbic.portal.offermanager.components.createaffiliation.CreateAffiliationPresenter
-import life.qbic.portal.offermanager.components.createperson.CreatePersonPresenter
-import life.qbic.portal.offermanager.components.createoffer.CreateOfferPresenter
+import life.qbic.portal.offermanager.components.affiliation.create.CreateAffiliationPresenter
+import life.qbic.portal.offermanager.components.person.create.CreatePersonPresenter
+import life.qbic.portal.offermanager.components.offer.create.CreateOfferPresenter
 import life.qbic.portal.offermanager.web.presenters.ListAffiliationsPresenter
 import life.qbic.portal.offermanager.components.AppPresenter
 import life.qbic.portal.offermanager.web.presenters.SearchCustomerPresenter
-import life.qbic.portal.offermanager.components.createaffiliation.CreateAffiliationViewModel
-import life.qbic.portal.offermanager.components.createperson.CreatePersonViewModel
-import life.qbic.portal.offermanager.components.createoffer.CreateOfferViewModel
-import life.qbic.portal.offermanager.components.offeroverview.OfferOverviewModel
+import life.qbic.portal.offermanager.components.affiliation.create.CreateAffiliationViewModel
+import life.qbic.portal.offermanager.components.person.create.CreatePersonViewModel
+import life.qbic.portal.offermanager.components.offer.create.CreateOfferViewModel
+import life.qbic.portal.offermanager.components.offer.overview.OfferOverviewModel
 import life.qbic.portal.offermanager.web.viewmodel.SearchCustomerViewModel
-import life.qbic.portal.offermanager.components.updateoffer.UpdateOfferViewModel
+import life.qbic.portal.offermanager.components.offer.update.UpdateOfferViewModel
 import life.qbic.portal.offermanager.components.AppViewModel
-import life.qbic.portal.offermanager.components.createaffiliation.CreateAffiliationView
-import life.qbic.portal.offermanager.components.createperson.CreatePersonView
-import life.qbic.portal.offermanager.components.createoffer.CreateOfferView
-import life.qbic.portal.offermanager.components.offeroverview.OfferOverviewView
+import life.qbic.portal.offermanager.components.affiliation.create.CreateAffiliationView
+import life.qbic.portal.offermanager.components.person.create.CreatePersonView
+import life.qbic.portal.offermanager.components.offer.create.CreateOfferView
+import life.qbic.portal.offermanager.components.offer.overview.OfferOverviewView
 import life.qbic.portal.offermanager.components.AppView
 import life.qbic.portal.offermanager.web.views.SearchCustomerView
 import life.qbic.portal.utils.ConfigurationManager
