+++ offer-manager-app/src/main/groovy/life/qbic/portal/offermanager/components/affiliation/create/CreateAffiliationViewModel.groovy
@@ -1,4 +1,4 @@
-package life.qbic.portal.offermanager.components.createaffiliation
+package life.qbic.portal.offermanager.components.affiliation.create
 
 import groovy.beans.Bindable
 import life.qbic.portal.offermanager.dataresources.ResourcesService
