+++ offer-manager-app/src/main/groovy/life/qbic/portal/offermanager/components/offer/create/CustomerSelectionView.groovy
@@ -1,4 +1,4 @@
-package life.qbic.portal.offermanager.components.createoffer.customerselection
+package life.qbic.portal.offermanager.components.offer.create
 
 
 import com.vaadin.data.provider.ListDataProvider
@@ -14,7 +14,7 @@ import com.vaadin.ui.themes.ValoTheme
 import life.qbic.datamodel.dtos.business.AcademicTitle
 import life.qbic.datamodel.dtos.business.Affiliation
 import life.qbic.datamodel.dtos.business.Customer
-import life.qbic.portal.offermanager.components.createoffer.CreateOfferViewModel
+import life.qbic.portal.offermanager.components.offer.create.CreateOfferViewModel
 import life.qbic.portal.offermanager.components.GridUtils
 
 /**
