+++ offer-manager-app/src/main/groovy/life/qbic/portal/offermanager/components/offer/update/UpdateOfferViewModel.groovy
@@ -1,17 +1,17 @@
-package life.qbic.portal.offermanager.components.updateoffer
+package life.qbic.portal.offermanager.components.offer.update
 
 import life.qbic.datamodel.dtos.business.Offer
-import life.qbic.portal.offermanager.components.createoffer.CreateOfferViewModel
+import life.qbic.portal.offermanager.components.offer.create.CreateOfferViewModel
 import life.qbic.portal.offermanager.dataresources.customers.PersonResourcesService
 import life.qbic.portal.offermanager.dataresources.products.ProductsResourcesService
 import life.qbic.portal.offermanager.dataresources.offers.OfferUpdateService
-import life.qbic.portal.offermanager.web.viewmodel.ProductItemViewModel
+import life.qbic.portal.offermanager.components.offer.create.ProductItemViewModel
 
 
 /**
  * Model with data for updating an existing offer.
  *
- * This specialisation of the {@link life.qbic.portal.offermanager.components.createoffer.CreateOfferViewModel} updates its data upon a received
+ * This specialisation of the {@link CreateOfferViewModel} updates its data upon a received
  * offer update event.
  *
  * With respect to its parent class, it contains an additional service and subscribes to an
