+++ offer-manager-app/src/main/groovy/life/qbic/portal/offermanager/components/person/create/CreatePersonController.groovy
@@ -1,4 +1,4 @@
-package life.qbic.portal.offermanager.components.createperson
+package life.qbic.portal.offermanager.components.person.create
 
 import groovy.util.logging.Log4j2
 import life.qbic.datamodel.dtos.business.AcademicTitle
