+++ offer-manager-app/src/test/groovy/life/qbic/portal/qoffer2/web/controllers/CreateCustomerControllerSpec.groovy
@@ -5,7 +5,7 @@ import life.qbic.datamodel.dtos.business.AcademicTitleFactory
 import life.qbic.datamodel.dtos.business.Affiliation
 import life.qbic.datamodel.dtos.business.Customer
 import life.qbic.business.customers.create.CreateCustomerInput
-import life.qbic.portal.offermanager.components.createperson.CreatePersonController
+import life.qbic.portal.offermanager.components.person.create.CreatePersonController
 import spock.lang.Specification
 
 /**
